+++ util/cbfstool/common.h
@@ -178,8 +178,8 @@ int parse_bzImage_to_payload(const struct buffer *input,
 			     char *cmdline, enum cbfs_compression algo);
 int parse_flat_binary_to_payload(const struct buffer *input,
 				 struct buffer *output,
-				 uint32_t loadaddress,
-				 uint32_t entrypoint,
+				 uint64_t loadaddress,
+				 uint64_t entrypoint,
 				 enum cbfs_compression algo);
 /* cbfs-mkstage.c */
 int parse_elf_to_stage(const struct buffer *input, struct buffer *output,
